+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+- log:
+  - Introduce AutoPager context manager for auto-paging feature on TTY (#44).
+  - Add dependency on `rfl.core`.
+
 ## [1.5.0] - 2025-06-27
 
 ### Added